requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
beautifulsoup4==4.9.1
Flask==3.0.0
zope.interface==5.1.0
requests
tqdm
python-dateutil==2.8.1
